pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "contextualizer"
version = "0.1.0"
description = "System tray app for managing work contexts with a local HTTP API"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pystray>=0.19",
    "Pillow>=10.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-dateutil>=2.9",
    "platformdirs>=2.6",
]

[project.scripts]
contextualizer = "contextualizer.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/contextualizer"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
